add deleteUser by id

+++ ms-andrian-betest/src/controllers/index.js
@@ -138,7 +138,6 @@ class Controller {
   static async deleteUserById(req, res) {
     const db = getDB();
     const userId = req.params._id;
-
     if (!ObjectId.isValid(userId)) {
       return res.status(400).json({
         message: 'Invalid ID format',
@@ -146,15 +145,20 @@ class Controller {
     }
 
     try {
+      const user = await db
+        .collection('users')
+        .findOne({ _id: new ObjectId(userId) });
+
+      if (!user) {
+        return res.status(404).json({ message: 'User  Not Found' });
+      }
       const result = await db
         .collection('users')
         .deleteOne({ _id: new ObjectId(userId) });
 
-      if (result.deletedCount === 0) {
-        return res.status(404).json({ message: 'User  Not Found' });
-      }
-
-      res.status(200).json({ message: 'User  deleted successfully' });
+      res
+        .status(200)
+        .json({ message: 'User  deleted successfully', user: user });
     } catch (err) {
       console.error(err);
       res.status(500).json({ error: 'Internal Server Error', details: err });
